Web-UI/requirements.txt
flask
sqlalchemy
bcrypt
python-magic
reportlab
